fix compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3222



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@14007 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/tools/toolutil/ucmstate.c
@@ -1018,7 +1018,7 @@ ucm_countChars(UCMStates *states,
                 break;
             default:
                 /* reserved, must never occur */
-                fprintf(stderr, "ucm error: byte sequence reached reserved action code, entry: 0x%02lx\n", entry);
+                fprintf(stderr, "ucm error: byte sequence reached reserved action code, entry: 0x%02lx\n", (unsigned long)entry);
                 return -1;
             }
         }
